refactor(session-history): split session-history.ts into per-layer submodules
Replace agent/session-history.ts (699 LOC) with a session-history/
directory that separates the layers the user identified
(parse / normalize / dedupe / metadata / IO):

- shared.ts      - types + constants + pure helpers (textFromContent,
                   thinkingFromContent, isChatRole, hasA2ui, trimText,
                   normalizeCwd)
- parse-pi.ts    - parseSessionHistoryLines (Anthropic JSONL transcript
                   with tool-card backfill)
- parse-local.ts - parseLocalChatLines + readLocalChatTranscript
                   (aethon-chat.jsonl append-log)
- io.ts          - normalizeSessionLabel, readSessionLabel,
                   writeSessionLabel, appendLocalChatMessage (with
                   atomic pruneLocalChatFile)
- metadata.ts    - latestSessionLog, readSessionMetadata (header cwd +
                   first-user-message extraction for sidebar labels)
- lookup.ts      - findSessionFileMatchingCwd + private
                   readSessionHeaderCwd (project-safety bridge for
                   ensureTab on the shared `default` dir)
- restore.ts     - readSessionTranscript + private dedupeLocalMessages
                   (orchestration over parse-pi + parse-local + lookup)
- index.ts       - barrel re-exporting the full public surface

Public surface unchanged. All 6 external consumers
(extension-loader.ts, main.ts, nativeSlash.ts, tabs.ts,
tab-lifecycle/lifecycle.ts, session-history.test.ts) keep importing
from "./session-history" which now resolves to the directory's
index.ts. The 27-test suite in session-history.test.ts stays green
(including the project-safety findSessionFileMatchingCwd cases).
